# postman/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BullBearBroker API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/markets/crypto/prices": {
      "get": {
        "tags": [
          "markets",
          "Markets"
        ],
        "summary": "Get Crypto Prices",
        "description": "Return crypto prices for the provided symbols.",
        "operationId": "get_crypto_prices_api_markets_crypto_prices_get",
        "parameters": [
          {
            "name": "symbols",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Lista de s\u00edmbolos separados por coma",
              "title": "Symbols"
            },
            "description": "Lista de s\u00edmbolos separados por coma"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Crypto Prices Api Markets Crypto Prices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/markets/stocks/quotes": {
      "get": {
        "tags": [
          "markets",
          "Markets"
        ],
        "summary": "Get Stock Quotes",
        "description": "Return stock prices for the provided tickers.",
        "operationId": "get_stock_quotes_api_markets_stocks_quotes_get",
        "parameters": [
          {
            "name": "symbols",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Lista de tickers separados por coma",
              "title": "Symbols"
            },
            "description": "Lista de tickers separados por coma"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Stock Quotes Api Markets Stocks Quotes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/markets/forex/rates": {
      "get": {
        "tags": [
          "markets",
          "Markets"
        ],
        "summary": "Get Forex Rates",
        "description": "Return forex rates for the given currency pairs.",
        "operationId": "get_forex_rates_api_markets_forex_rates_get",
        "parameters": [
          {
            "name": "pairs",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Pares FX separados por coma",
              "title": "Pairs"
            },
            "description": "Pares FX separados por coma"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Forex Rates Api Markets Forex Rates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/markets/crypto/{symbol}": {
      "get": {
        "tags": [
          "markets",
          "Markets"
        ],
        "summary": "Get Crypto",
        "description": "Retrieve crypto pricing information ensuring service fallback order.",
        "operationId": "get_crypto_api_markets_crypto__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Crypto Api Markets Crypto  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/markets/history/{symbol}": {
      "get": {
        "tags": [
          "markets",
          "Markets"
        ],
        "summary": "Get History",
        "operationId": "get_history_api_markets_history__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1h",
              "title": "Interval"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 10,
              "default": 300,
              "title": "Limit"
            }
          },
          {
            "name": "market",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(auto|crypto|stock|equity|forex)$",
              "default": "auto",
              "title": "Market"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get History Api Markets History  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/markets/stock/{symbol}": {
      "get": {
        "tags": [
          "markets",
          "Markets"
        ],
        "summary": "Get Stock",
        "description": "Retrieve stock pricing information delegating to the StockService cascade.",
        "operationId": "get_stock_api_markets_stock__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Stock Api Markets Stock  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/markets/forex/{pair}": {
      "get": {
        "tags": [
          "markets",
          "Markets"
        ],
        "summary": "Get Forex",
        "description": "Retrieve forex quote information honouring the configured fallback chain.",
        "operationId": "get_forex_api_markets_forex__pair__get",
        "parameters": [
          {
            "name": "pair",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pair"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Forex Api Markets Forex  Pair  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}